requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiodns==3.1.1
    # via aiohttp
aiohttp[speedups]==3.8.6
    # via py-cord
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.2.0
    # via aiohttp
brotli==1.1.0
    # via aiohttp
cffi==1.16.0
    # via pycares
charset-normalizer==3.3.2
    # via aiohttp
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
idna==3.6
    # via yarl
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
orjson==3.9.10
    # via py-cord
py-cord[speed]==2.4.1
    # via -r requirements.in
pycares==4.4.0
    # via aiodns
pycparser==2.21
    # via cffi
python-dotenv==1.0.0
    # via -r requirements.in
typing-extensions==4.9.0
    # via py-cord
yarl==1.9.4
    # via aiohttp
